+++ src/shared/agent-title-evidence.ts
@@ -1,18 +1,21 @@
 import {
-  AGY_AGENT_NAME_RE,
   CLAUDE_IDLE,
-  DROID_AGENT_NAME_RE,
   GEMINI_IDLE,
   GEMINI_PERMISSION,
   GEMINI_SILENT_WORKING,
   GEMINI_WORKING,
-  HERMES_AGENT_NAME_RE,
   containsAgentSpinnerGlyph,
   isClaudeIdentityFrameSegment,
   isClaudeManagementTitle,
-  isCursorNativeAgentTitle,
-  titleHasAgentName
+  isCursorNativeAgentTitle
 } from './agent-title-core'
+import {
+  DISPLAY_LABELS,
+  agentForBareName,
+  agentForWholeTitle,
+  namesIn,
+  stripBareNameDecoration
+} from './agent-title-name-anchoring'
 import { isOpenCodeNativeTitle } from './opencode-terminal-title'
 import { stripLeadingAgentTitleDecorationOrEmpty } from './agent-title-decoration'
 import { getPiCompatibleSyntheticAgentLabel } from './pi-compatible-synthetic-title'
@@ -21,7 +24,6 @@ import {
   SYNTHETIC_AGENT_TITLE_PROFILES
 } from './synthetic-agent-title'
 import type { TuiAgent } from './tui-agent'
-import { TUI_AGENT_DISPLAY_NAMES } from './tui-agent-display-names'
 
 /**
  * Order-independent identity evidence from a terminal title.
@@ -57,48 +59,6 @@ export type AgentTitleEvidence = {
   readonly reason: AgentTitleEvidenceReason
 }
 
-/** Names matched as whole tokens, paired with the agent each identifies. */
-const NAME_TOKENS: readonly (readonly [string, TuiAgent])[] = [
-  ['claude', 'claude'],
-  ['openclaude', 'openclaude'],
-  ['codex', 'codex'],
-  ['copilot', 'copilot'],
-  ['cursor', 'cursor'],
-  ['gemini', 'gemini'],
-  ['antigravity', 'antigravity'],
-  ['opencode', 'opencode'],
-  ['mimo', 'mimo-code'],
-  ['openclaw', 'openclaw'],
-  ['aider', 'aider'],
-  ['grok', 'grok'],
-  ['devin', 'devin']
-]
-
-/** Agents whose name is matched by a dedicated pattern rather than a plain token. */
-const PATTERN_NAMES: readonly (readonly [RegExp, TuiAgent])[] = [
-  [AGY_AGENT_NAME_RE, 'antigravity'],
-  [DROID_AGENT_NAME_RE, 'droid'],
-  [HERMES_AGENT_NAME_RE, 'hermes']
-]
-
-/** Catalog labels known to be emitted as terminal titles, not merely presented in Orca's UI. */
-const EMITTED_DISPLAY_LABEL_AGENTS = [
-  'claude-agent-teams',
-  'mimo-code',
-  'prime-agent',
-  'command-code',
-  'copilot'
-] as const satisfies readonly TuiAgent[]
-
-const DISPLAY_LABELS = [
-  ...EMITTED_DISPLAY_LABEL_AGENTS.map(
-    (agent) => [TUI_AGENT_DISPLAY_NAMES[agent].toLowerCase(), agent] as const
-  ),
-  ['claude code', 'claude'],
-  ['gemini cli', 'gemini'],
-  ['agent teams', 'claude-agent-teams']
-] satisfies readonly (readonly [string, TuiAgent])[]
-
 const GEMINI_GLYPHS = [GEMINI_WORKING, GEMINI_SILENT_WORKING, GEMINI_IDLE, GEMINI_PERMISSION]
 const ANTIGRAVITY_MODEL_TITLE_RE = /^(?:agy|antigravity)(?:\s*[·—:-]\s*|\s+)gemini\s+\d/i
 
@@ -108,7 +68,6 @@ const ANTIGRAVITY_MODEL_TITLE_RE = /^(?:agy|antigravity)(?:\s*[·—:-]\s*|\s+)g
  * worktree name (`review-14600-codex`), which is a directory, not an owner declaration.
  */
 const OWNER_SUFFIX_RE = /\s-\s+([A-Za-z][\w-]*)\s*$/
-const WINDOWS_LAUNCHER_SUFFIX_RE = /\.(?:exe|cmd|bat|ps1)$/i
 const WRAPPER_SEPARATOR = ' | '
 const MAX_WRAPPER_EVIDENCE_SEGMENTS = 8
 const RESERVED_OWNER_IDS: ReadonlyMap<string, TuiAgent> = new Map([
@@ -133,63 +92,6 @@ function getEvidenceTitleSegments(title: string): string[] {
     }
   }
   return segments
-}
-
-function namesIn(text: string): TuiAgent[] {
-  const found = new Set<TuiAgent>()
-  for (const [token, agent] of NAME_TOKENS) {
-    if (titleHasAgentName(text, token)) {
-      found.add(agent)
-    }
-  }
-  for (const [pattern, agent] of PATTERN_NAMES) {
-    if (pattern.test(text)) {
-      found.add(agent)
-    }
-  }
-  return [...found]
-}
-
-function stripBareNameDecoration(text: string): string {
-  return text
-    .trim()
-    .replace(/^[^\p{L}\p{N}]+/u, '')
-    .replace(/[^\p{L}\p{N}]+$/u, '')
-}
-
-function agentForBareName(text: string): TuiAgent | null {
-  const trimmed = text.trim()
-  if (!trimmed || /[\\/]/.test(trimmed)) {
-    return null
-  }
-  const stripped = stripBareNameDecoration(trimmed)
-  // Why labels too: an agent may write its own display name as the entire title (`⠐ Claude Code`).
-  // That is the same claim as a bare token, just spelled the way the vendor spells it.
-  const label = DISPLAY_LABELS.find(([text]) => text === stripped.toLowerCase())
-  if (label) {
-    return label[1]
-  }
-  const bareToken = stripped.replace(WINDOWS_LAUNCHER_SUFFIX_RE, '')
-  const names = namesIn(bareToken)
-  // Why the length check: the remainder must BE the name, not merely contain it. "agy" anchors;
-  // "fix the agy hook" does not, and neither does a hyphenated worktree name like "codex-split".
-  return names.length === 1 && /^[\p{L}\p{N}]+$/u.test(bareToken) ? names[0] : null
-}
-
-function agentForWholeTitle(text: string): TuiAgent | null {
-  const trimmed = text.trim()
-  if (!trimmed || /[\\/]/.test(trimmed)) {
-    return null
-  }
-  const stripped = stripBareNameDecoration(trimmed)
-  const label = DISPLAY_LABELS.find(([text]) => text === stripped.toLowerCase())
-  if (label) {
-    return label[1]
-  }
-  if (!WINDOWS_LAUNCHER_SUFFIX_RE.test(stripped)) {
-    return null
-  }
-  return agentForBareName(stripped)
 }
 
 function agentForOwnerSuffix(text: string): TuiAgent | null {
